Continued refactoring in aprgcommon module

+++ AprgCommon/AprgCommon/src/Common/String/AlbaStringHelper.cpp
@@ -521,15 +521,13 @@ string getHexNumberAfterThisString(string_view mainString, string_view stringToS
 string getHexEquivalentOfCharacters(string_view stringToCheck) {
     stringstream ss;
     for (unsigned char const c : stringToCheck) {
-        ss << hex << uppercase << setfill('0') << setw(2) << static_cast<unsigned int>(c);
+        ss << hex << uppercase << setfill('0') << setw(2) << static_cast<unsigned long long>(c);
     }
     return ss.str();
 }
-
 string constructFileLocator(string_view file, int const lineNumber) {
     stringstream ss;
-    ss << file.substr(file.find_last_of('\\') + 1) << "[" << lineNumber << "]";
-    return ss.str();
+    ss << file.substr(file.find_last_of('\\') + 1) << "[" << lineNumber << "]";    return ss.str();
 }
 
 string getRandomAlphaNumericString(size_t const length) {
