+++ AprgCommon/AprgCommon/tst/CommonTests/Randomizer/AlbaOldRandomizer_unit.cpp
@@ -26,19 +26,17 @@ TEST(AlbaOldRandomizerTest, GetRandomIntegerInUniformDistributionWorks_AsUniform
     int numberOfRandomValues(maximumValue - minimumValue + 1);
     int const iterations(1000);
     int const allowedDeviation(iterations / 10);
-    vector<int> hitsForEachValue(static_cast<unsigned long>(numberOfRandomValues), 0);
+    vector<int> hitsForEachValue(numberOfRandomValues, 0);
 
     for (int i = 0; i < iterations; i++) {
         int random(randomizer.getRandomIntegerInUniformDistribution(minimumValue, maximumValue));
-        hitsForEachValue[static_cast<unsigned long>(random)]++;
+        hitsForEachValue[random]++;
     }
 
-    ASSERT_FALSE(hitsForEachValue.empty());
-    auto&& [maxIterator, minIterator] = std::minmax_element(hitsForEachValue.cbegin(), hitsForEachValue.cend());
+    ASSERT_FALSE(hitsForEachValue.empty());    auto&& [maxIterator, minIterator] = std::minmax_element(hitsForEachValue.cbegin(), hitsForEachValue.cend());
     int deviation(*maxIterator - *minIterator);
     EXPECT_LE(deviation, allowedDeviation);
 }
-
 TEST(AlbaOldRandomizerTest, GetRandomValueInUniformDistributionWorks_WithinMinimumAndMaximumValues) {
     AlbaOldRandomizer randomizer;
     constexpr double minimumValue(-11.5);
